Fix OFDM test build: separate compilation, include guards, #include_next
Switch from direct .c inclusion to separate object compilation to avoid
re-declaration conflicts from auto-generated headers that lack include
guards. Each source (ofdm_tx.c, ofdm_rx.c, cw_tone.c, timers.c) is
compiled independently with -Istubs and linked together.

- Add ofdm_test_shims.c with all stub functions as separate TU
- Add include-guarded wrapper headers in stubs/ for all guardless
  auto-generated headers (timers.h, pluto.h, charon.h, etc.)
- Fix stubs/fftw3.h infinite recursion: use #include_next instead of
  #include <fftw3.h>
- Update Makefile with per-object build rules and ofdm_build/ directory
