+++ tests/stubs/charon.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../charon.h (which lacks a guard)
+#ifndef STUB_CHARON_H
+#define STUB_CHARON_H
+#include "../../charon.h"
+#endif
