+++ tests/stubs/config.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../config.h (which lacks a guard)
+#ifndef STUB_CONFIG_H
+#define STUB_CONFIG_H
+#include "../../config.h"
+#endif
