+++ tests/stubs/crc.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../crc.h (which lacks a guard)
+#ifndef STUB_CRC_H
+#define STUB_CRC_H
+#include "../../crc.h"
+#endif
