+++ tests/stubs/cw_tone.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../cw_tone.h (which lacks a guard)
+#ifndef STUB_CW_TONE_H
+#define STUB_CW_TONE_H
+#include "../../cw_tone.h"
+#endif
