+++ tests/stubs/fftw3.h
@@ -1,3 +1,5 @@
 // Redirect quoted "fftw3.h" include (used by ofdm_rx.c) to the real
-// system header installed by libfftw3-dev.
-#include <fftw3.h>
+// system header installed by libfftw3-dev.  We use #include_next so
+// that -Istubs doesn't cause infinite self-inclusion.
+#pragma once
+#include_next <fftw3.h>
