+++ tests/stubs/ofdm_rx.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../ofdm_rx.h (which lacks a guard)
+#ifndef STUB_OFDM_RX_H
+#define STUB_OFDM_RX_H
+#include "../../ofdm_rx.h"
+#endif
