+++ tests/stubs/ofdm_tx.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../ofdm_tx.h (which lacks a guard)
+#ifndef STUB_OFDM_TX_H
+#define STUB_OFDM_TX_H
+#include "../../ofdm_tx.h"
+#endif
