+++ tests/stubs/pluto.h
@@ -1,0 +1,6 @@
+// Include-guarded wrapper for ../pluto.h (which lacks a guard)
+// Functions are stubbed in test_ofdm.c — only declarations needed here.
+#ifndef STUB_PLUTO_H
+#define STUB_PLUTO_H
+#include "../../pluto.h"
+#endif
