+++ tests/stubs/tap_device.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../tap_device.h (which lacks a guard)
+#ifndef STUB_TAP_DEVICE_H
+#define STUB_TAP_DEVICE_H
+#include "../../tap_device.h"
+#endif
