+++ tests/stubs/timers.h
@@ -1,0 +1,5 @@
+// Include-guarded wrapper for ../timers.h (which lacks a guard)
+#ifndef STUB_TIMERS_H
+#define STUB_TIMERS_H
+#include "../../timers.h"
+#endif
